Adding special thanks and code snippet reference
Thank you so much for my mentor and the original author of the code snippets being referenced. We looked up so many pages for the count-down animation and got inspirations from so many of them. I apologize for not listing up every single page that we saw on here, but I truly appreciate all the authors who selflessly opened up their own code to the internet and let many benefited from learning from their examples. Lots lots of thanks!